Extracts LoggingFactory into an interface.
A backward compatible DefaultLoggerFactory is included.
