+++ dropwizard-core/src/main/java/io/dropwizard/Configuration.java
@@ -2,6 +2,7 @@ package io.dropwizard;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Objects;
+import io.dropwizard.logging.DefaultLoggingFactory;
 import io.dropwizard.logging.LoggingFactory;
 import io.dropwizard.metrics.MetricsFactory;
 import io.dropwizard.server.DefaultServerFactory;
@@ -65,7 +66,7 @@ public class Configuration {
 
     @Valid
     @NotNull
-    private LoggingFactory logging = new LoggingFactory();
+    private LoggingFactory logging = new DefaultLoggingFactory();
 
     @Valid
     @NotNull
