+++ dropwizard-jersey/src/test/java/io/dropwizard/jersey/sessions/FlashProviderTest.java
@@ -5,7 +5,7 @@ import com.sun.jersey.api.client.WebResource;
 import com.sun.jersey.test.framework.AppDescriptor;
 import com.sun.jersey.test.framework.JerseyTest;
 import com.sun.jersey.test.framework.WebAppDescriptor;
-import io.dropwizard.logging.LoggingFactory;
+import io.dropwizard.logging.DefaultLoggingFactory;
 import org.junit.Test;
 
 import javax.ws.rs.core.MediaType;
@@ -16,7 +16,7 @@ import static org.fest.assertions.api.Assertions.assertThat;
 
 public class FlashProviderTest extends JerseyTest {
     static {
-        LoggingFactory.bootstrap();
+        DefaultLoggingFactory.bootstrap();
     }
 
     @Override
